requirements.txt
amqp==5.1.1
APScheduler==3.6.3
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
Brotli==1.0.9
cachetools==4.2.2
celery==5.2.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
dash==2.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Deprecated==1.2.13
Django==4.0.4
django-celery-beat==2.3.0
django-timezone-field==5.0
docxtpl==0.16.0
Flask==2.1.2
Flask-Compress==1.12
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib==3.5.2
packaging==21.3
plotly==5.8.0
Pillow==9.1.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
pyparsing==3.0.9
pyTelegramBotAPI==4.5.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-telegram-bot==13.12
python-telegram-bot-raw==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
telebot==0.0.4
tenacity==8.0.1
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
